Add files via upload
Scripts to populate a cubic lattice with a Higgs field distribution by randomly picking t'Hoft angles. Modules include routines to compute magnetic field directly and the gauge potential directly and the curl, magnetic field spectrum according to the prescription in Yiyang et al. Experimental plotting scripts to plot enseble averaged and binned spectra are included along with sample plots from 120 runs run of (128+6)^3 lattices. +6 is required for the ghost zones in the pencil-code.